+++ src/tailwind/form/LabelledSelect.tsx
@@ -1,11 +1,11 @@
-import type { FC } from 'react';
+import type { FC, ReactNode } from 'react';
 import { useId } from 'react';
 import { Label } from './Label';
 import type { SelectProps } from './Select';
 import { Select } from './Select';
 
 export type LabelledSelectProps = Omit<SelectProps, 'className' | 'id'> & {
-  label: string;
+  label: ReactNode;
   selectClassName?: string;
 
   /** Alternative to `required`. Causes the input to be required, without displaying an asterisk */